requirements.txt
python-chess
openai
python-dotenv